// src/app/models/product.ts
export class Product {
    ProductId: number;
    CatId: number;
    ResId: number;
    Price: number;
    OfferId: number;
    DeliveryTime: string;
    Extra: string;
    IsActive: boolean;
    Remark: string;
    ImageName: string;
    ImagePath: string;
    StateId: number;
    CityId: number;
    Name: string;   
    Qty: string; 
    Amount: number;      
}
													